Update README.md
Taskmate is a task management web application built with the Django Framework. The goal of this project was to study the working of the Django Framework, including its common features such as MVT architecture, working with forms, default authentication,.env files, and the deployment procedure on the Railway platform with PostgreSQL database.